Upgrade to Babel 6
Biggest changes:
1. Including all the various plugins that Babel 5 came with by default
2. Moving the module plugin config where it now belongs

(1) was mechanical, but I'm a Babel n00b, so please take a careful look at
that. I just copied what @kittens did for React +flow +object-rest-spread.

For (2) I changed the default-options module to export a function that
takes a bit of configuration. The module takes care of threading the
plugin options to the right place. This seemed better than deeply
mutating a static export.

Next step here is to apply the module rewrite plugin to .js.flow files,
which we should be able to do with Babel 6.

Test Plan: I compared the diff between the existing lib output and the
new lib output. The `"use strict";` pragma moved under the first
comment, but that seems fine. Some extra parens. Seems fine.