Relativize requires in .flow.js files
Keep `@providesModule` and workaround Babel bug

The babel bug is a fun one. `?T` gets becomes `?TT`, which can't be
resolved. I found this when testing relay for the `@providesModule`
thing. `T|void|null` is equivalent to `?T`, but I'm going to put up a PR
for Babel in moments...

Add babel-plugin-syntax-flow