Add old Promise Flow definition
Flow 0.23 removed the un-speced methods from the Promise class, which is
the right thing to do. However our Promise impl does have those methods
and we use them so for now, the best thing to do is just ship the
Promise class definition that we are using. In the future we can stop
relying on that method.

Also fixed gitignore to pick up changes in `flow/lib` properly.