Bump "promise" to ^8.0.3 to match React Native
React Native 0.63 (and by extension Expo SDK 39-40) bumped its dependency
on "promise" to be ^8.0.3 ¹

This repo depends on promise@^7.1.1 which has no valid overlapping versions. This module
"fbjs" is depended on by "metro", which in turn is depended on by RN/Expo projects.

This means that on a standard RN/Expo project, two versions of "promise" get installed,
making it impossible to obtain a reference to the same version that React Native is using²,
and thus impossible to detect unhandled promise rejections – and for example report them
to Bugsnag.

¹ https://github.com/facebook/react-native/commit/b23efc526446cad53adc3de7465fd8be7309d84e
² https://github.com/bugsnag/bugsnag-js/blob/eb6e05f350f5acf496c9e3dfd341c9061c400c1f/packages/plugin-react-native-unhandled-rejection/rejection-handler.js#L5